Rename `--registry-poll-interval`
The flag `--registry-poll-interval` used to control how often fluxd
would scan image repos for new images, and consequently update any
automated workloads. This has not been true for some time: it now
just controls the second part, while images are scanned on a
different schedule. The name is a persistent source of confusion.

Therefore: introduce a new flag `--automation-interval`, and make
`--registry-poll-interval` a synonym so as not to break things
suddenly, and deprecate the latter so as to encourage people to
change it.